v0.8.1: Complete cross-compiler build system with strict warning support
Major build system enhancement for comprehensive compiler compatibility:

🔧 BUILD SYSTEM ENHANCEMENTS:
• MSVCStrict/MSVCWarn: Full MSVC-like strict warning modes for Clang
• GCCStrict/GCCWarn: Complete GCC strict warning and diagnostic modes
• Cross-compiler compatibility: Clang, GCC, MSVC with consistent behavior
• Configurable warning-as-error vs warning-only modes for development
• Architecture-specific SIMD detection and safe compilation flags

📋 WARNING LEVEL STANDARDIZATION:
• Type conversion warnings (-Wconversion, -Wfloat-conversion, etc.)
• Old-style cast detection (-Wold-style-cast)
• Implicit conversion detection (-Wimplicit-int-conversion)
• GCC-specific diagnostics (logical-op, duplicated-cond, etc.)
• MSVC-equivalent strictness levels across all compilers

🚀 PRACTICAL OUTCOMES:
• Zero-warning builds across all supported compilers and modes
• Production-ready strict compilation without compatibility issues
• Developer-friendly warning modes that don't block iteration
• Consistent behavior regardless of target compiler environment

This enhances libstats' as a professional build standards supporting
diverse compiler toolchains and CI/CD pipelines.

+++ examples/basic_usage.cpp
@@ -169,7 +169,7 @@ int main() {
     std::cout << std::endl;
     
     std::cout << "\n   ⚡ Performance: Processed 16 PDFs in " << duration.count() << " nanoseconds"
-              << " (" << std::setprecision(2) << duration.count() / 16.0 << " ns per PDF)" << std::endl;
+              << " (" << std::setprecision(2) << static_cast<double>(duration.count()) / 16.0 << " ns per PDF)" << std::endl;
     std::cout << "   ✓ libstats selected optimal strategy based on data size and system capabilities" << std::endl;
     
     std::cout << "=== Example completed successfully ===" << std::endl;
