+++ include/platform/platform_constants.h
@@ -826,7 +826,7 @@ namespace platform {
             return base_grain * 2;  // Can handle larger chunks efficiently
         }
         else if (features.avx2 || features.fma) {
-            return static_cast<std::size_t>(std::round(base_grain * 1.5));
+            return static_cast<std::size_t>(std::round(static_cast<double>(base_grain) * 1.5));
         }
         else {
             return base_grain;
